♻️ workspaceをuseTechnologyに分離
