+++ app/javascript/components/Workspace.js
@@ -4,15 +4,13 @@ import Technology from "./Technology";
 import Arrow from "./Arrow";
 import useArrow from "../hooks/useArrow";
 import useComponentArray from "../hooks/useComponentArray";
+import useTechnology from '../hooks/useTechnology';
 
 const Workspace = (props) => {
-  const [selectedNodes, setSelectedNodes] = useState([])
-  const [topTechnology, setTopTechnology] = useState(null)
   const [arrows, setArrows] = useState([])
   const [getArrowProperties] = useArrow(null)
   const technologyRefs = useRef([])
-  const beforeMovePosition = useRef()
-  const afterMovePosition = useRef()
+  const [topTechnology, setTopTechnology, selectedNodes, setSelectedNodes, liftUpTechnology, liftDown, selectTechnology] = useTechnology()
 
   const arrowHandler = () => {
     const childElements = selectedNodes.map((selectedId) => {
@@ -22,52 +20,6 @@ const Workspace = (props) => {
       const arrowProperties = getArrowProperties(topTechnology, element.current)
       setArrows(prevArrows => [...prevArrows, <Arrow width={arrowProperties.rect.width} height={arrowProperties.rect.height} startPosition={[arrowProperties.positions.startPosition.x, arrowProperties.positions.startPosition.y]}/>])
     })
-  }
-
-  const mousePosition = (evt) => {
-    return {x: evt.pageX, y: evt.pageY}
-  }
-
-  const haveNode = (evt) => {
-      const techElement = evt.target
-      techElement.style.position = 'absolute'
-      beforeMovePosition.current = mousePosition(evt)
-      console.log("beforeMovePosition", beforeMovePosition)
-  }
-
-  const dragEndHandler = (evt) => {
-    afterMovePosition.current = mousePosition(evt)
-    console.log('beforeMovePosition',beforeMovePosition.current)
-    const differentialMousePosX = afterMovePosition.current.x - beforeMovePosition.current.x
-    const differentialMousePosY = afterMovePosition.current.y - beforeMovePosition.current.y
-    evt.target.style.left = (evt.pageX - evt.nativeEvent.offsetX + differentialMousePosX) + "px"
-    evt.target.style.top = (evt.pageY - evt.nativeEvent.offsetY + differentialMousePosY) + "px"
-  }
-
-  const selectHandler = (e) => {
-    //TODO:topTechnologyHandlerと順序依存関係あるため要修正
-    if (e.ctrlKey) {
-      if (!selectedNodes.includes(e.target.id)) {
-        setSelectedNodes([...selectedNodes, e.target.id])
-      }
-    }
-    else if(e.keydown == null) {
-      setSelectedNodes([e.target.id])
-      setTopTechnology(null)
-    }
-  }
-
-  const topTechnologyHandler = (e) => {
-    //TODO:selectHandlerと順序依存関係あるため要修正
-    if (e.shiftKey) {
-      setTopTechnology(e.target)
-      setSelectedNodes(selectedNodes.filter((value) => value !== e.target.id))
-    }
-  }
-
-  const technologyHandler = (e) => {
-    haveNode(e)
-    e.stopPropagation()
   }
 
   return(
@@ -85,14 +37,14 @@ const Workspace = (props) => {
         {props.technologies.map((technology, i) => {
           technologyRefs.current[i] = React.createRef()
           return(
-            <div draggable onMouseDown={(e) => {
-              selectHandler(e)
-              topTechnologyHandler(e)
-              e.stopPropagation()
-              }
-              } 
-              onDragStart={technologyHandler} onDragEnd={dragEndHandler}
-              >
+            <div draggable
+              onMouseDown={(e) => {
+                selectTechnology(e)
+                e.stopPropagation()
+              }}
+              onDragStart={liftUpTechnology}
+              onDragEnd={liftDown}
+            >
               <Technology
                 name={technology.name}
                 key={technology.id}
