+++ app/javascript/hooks/useComponentArray.js
